file restructure yaml file edit
