+++ src/Components/App/App.js
@@ -2,13 +2,12 @@ import './App.css';
 import React, { useEffect } from 'react';
 import { getAllWorldsData } from './apiCalls'; 
 import { useDispatch } from 'react-redux';
-import { getDiscoveredWorlds } from './app/rootSlice';
+import { getDiscoveredWorlds } from './Components/rootSlice';
 import { Route, Switch } from 'react-router-dom';
-import { SingleWorld } from './app/SingleWorld/SingleWorld';
-import { WelcomePage } from './app/WelcomePage/WelcomePage'
-import { Header } from './app/Header/Header.js'
-import { WorldBrowser } from './app/WorldBrowser/WorldBrowser';
-
+import { SingleWorld } from './Components/SingleWorld/SingleWorld';
+import { WelcomePage } from './Components/WelcomePage/WelcomePage'
+import { Header } from './Components/Header/Header.js'
+import { WorldBrowser } from './Components/WorldBrowser/WorldBrowser';
 function App() {
   const dispatch = useDispatch()
